perf(bench): isolate heavyweight benchmark dependencies
