+++ Cargo.toml
@@ -102,11 +102,7 @@ either = { version = "1.13", features = ["serde"] }
 etcd-client = { version = "0.17.0", features = ["tls"] }
 futures = { version = "0.3" }
 futures-util = { version = "0.3.32" }
-hf-hub = { version = "0.4.2", default-features = false, features = [
-    "tokio",
-    "rustls-tls",
-    "ureq",
-] }
+hf-hub = { version = "0.4.2", default-features = false }
 
 # ModelExpress for model downloading
 modelexpress-client = { version = "0.3.0" }
