+++ lib/bench/Cargo.toml
@@ -12,33 +12,78 @@ repository.workspace = true
 description = "Lightweight HTTP benchmarks for Dynamo endpoints"
 
 [features]
-default = []
+# Preserve the package's default tests; benchmark entrypoints use the
+# documented `--no-default-features` feature sets below.
+default = ["claude-trace-export", "test-support"]
+active-sequences = [
+  "router-bench",
+  "dep:dynamo-tokens",
+  "dep:libc",
+  "dep:rustc-hash",
+  "dep:tracing-subscriber",
+]
+claude-trace-export = [
+  "dep:chrono",
+  "dep:crossbeam-channel",
+  "dep:dynamo-data-gen",
+  "dep:hf-hub",
+  "dep:rustc-hash",
+  "dep:sha2",
+  "dep:tokenizers",
+]
 dc-ckf-bench = []
-mocker-kvbm-offload = ["dynamo-mocker/kvbm-offload"]
-profile = ["dynamo-mocker/profile"]
+mooncake = ["router-bench", "dep:libc", "dep:tokio-util"]
+multiturn = ["dep:futures-util", "dep:rand", "dep:reqwest"]
+offline-replay = []
+router-bench = ["dynamo-kv-router/bench"]
+router-metrics = ["dynamo-kv-router/metrics"]
+router-test-support = [
+  "router-bench",
+  "router-metrics",
+  "dep:dynamo-tokens",
+  "dep:libc",
+  "dep:rustc-hash",
+  "dep:tokio-util",
+  "dep:tracing-subscriber",
+]
+satf = ["dep:dynamo-data-gen"]
+test-support = [
+  "router-test-support",
+  "dep:flate2",
+  "dep:tempfile",
+  "dep:tracing",
+  "dep:uuid",
+]
+mocker-kvbm-offload = ["offline-replay", "dynamo-mocker/kvbm-offload"]
+profile = ["offline-replay", "dynamo-mocker/profile"]
 
 [[bench]]
 name = "multiturn_bench"
 path = "multiturn_bench.rs"
 harness = false
+required-features = ["multiturn"]
 
 [[bin]]
 name = "claude_trace_export"
 path = "claude_trace_export.rs"
+required-features = ["claude-trace-export"]
 
 [[bin]]
 name = "request_trace_to_satf"
 path = "request_trace_to_satf.rs"
+required-features = ["satf"]
 
 [[bench]]
 name = "offline_replay_bench"
 path = "offline_replay_bench.rs"
 harness = false
+required-features = ["offline-replay"]
 
 [[bench]]
 name = "mooncake_bench"
 path = "kv_router/mooncake_bench.rs"
 harness = false
+required-features = ["mooncake"]
 
 [[bin]]
 name = "dc_ckf_consumer_bench"
@@ -49,6 +94,7 @@ required-features = ["dc-ckf-bench"]
 name = "active_sequences_bench"
 path = "kv_router/active_sequences_bench.rs"
 harness = false
+required-features = ["active-sequences"]
 
 [[bench]]
 name = "dc_ckf_relay_bench"
@@ -58,39 +104,32 @@ required-features = ["dc-ckf-bench"]
 
 [dependencies]
 anyhow = { workspace = true }
-chrono = { workspace = true }
+chrono = { workspace = true, optional = true }
 clap = { version = "4.5", features = ["derive"] }
-crossbeam-channel = "0.5"
-flate2 = "1"
-futures-util = { workspace = true }
-hf-hub = { workspace = true }
+crossbeam-channel = { version = "0.5", optional = true }
+flate2 = { version = "1", optional = true }
+futures-util = { workspace = true, optional = true }
+hf-hub = { workspace = true, features = ["ureq"], optional = true }
 indicatif = "0.18"
-rand = { workspace = true }
-reqwest = { workspace = true }
-rustc-hash = "2"
+rand = { workspace = true, optional = true }
+reqwest = { workspace = true, optional = true }
+rustc-hash = { version = "2", optional = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
-sha2 = "0.10"
+sha2 = { version = "0.10", optional = true }
+tempfile = { workspace = true, optional = true }
 tokio = { workspace = true }
+tokio-util = { workspace = true, optional = true }
 tokenizers = { version = "0.21.4", default-features = false, features = [
   "onig",
-  "esaxx_fast",
-  "rustls-tls",
-] }
-dynamo-tokens = { workspace = true }
-dynamo-data-gen = { workspace = true }
+], optional = true }
+tracing = { workspace = true, optional = true }
+tracing-subscriber = { workspace = true, optional = true }
+uuid = { workspace = true, optional = true }
+dynamo-tokens = { workspace = true, optional = true }
+dynamo-data-gen = { workspace = true, optional = true }
 dynamo-mocker = { workspace = true }
-dynamo-kv-router = { workspace = true, features = ["bench", "metrics", "shard-metrics"] }
+dynamo-kv-router = { workspace = true }
 
 [target.'cfg(target_os = "linux")'.dependencies]
-libc = { workspace = true }
-
-[dev-dependencies]
-dynamo-tokens = { workspace = true }
-minstant = "0.1.7"
-tempfile = { workspace = true }
-tokio = { workspace = true, features = ["rt", "macros", "time"] }
-tokio-util = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-uuid = { workspace = true }
+libc = { workspace = true, optional = true }
