+++ lib/bench/kv_router/common/replay.rs
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-use dynamo_kv_router::protocols::KvCacheEventData;
 #[allow(unused_imports)]
-pub use dynamo_kv_router::test_utils::NoopSequencePublisher;
+pub use dynamo_kv_router::NoopSequencePublisher;
+use dynamo_kv_router::protocols::KvCacheEventData;
 use dynamo_mocker::common::protocols::MockEngineArgs;
 use dynamo_mocker::loadgen::{SessionPartitionSpec, Trace};
 use dynamo_mocker::replay::ReplayKvEventVisibility;
