+++ lib/bench/src/lib.rs
@@ -1,7 +1,9 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+#[cfg(feature = "claude-trace-export")]
 pub mod coding;
+#[cfg(feature = "multiturn")]
 pub mod common;
 
 #[path = "../kv_router/common/mod.rs"]
