+++ lib/kv-router/Cargo.toml
@@ -18,7 +18,6 @@ metrics = ["dep:prometheus"]
 runtime-protocols = ["dep:dynamo-runtime"]
 bench = []
 profile = []
-shard-metrics = []
 standalone-indexer = ["dep:axum", "dep:serde_json", "dep:reqwest", "dep:zmq", "dep:chrono", "dep:tracing-appender"]
 standalone-slot-tracker = ["dep:axum", "dep:serde_json", "dep:zmq"]
 standalone-selection = ["standalone-indexer"]
