+++ lib/kv-router/src/test_utils.rs
@@ -4,15 +4,14 @@
 //! Shared test utilities for radix tree tests.
 
 use std::collections::HashSet;
-use std::future;
 
 use crate::indexer::KvIndexerInterface;
 use crate::protocols::{
-    ActiveLoad, ActiveSequenceEvent, ExternalSequenceBlockHash, KvCacheEvent, KvCacheEventData,
-    KvCacheRemoveData, KvCacheStoreData, KvCacheStoredBlockData, LocalBlockHash, OverlapScores,
-    RouterEvent, WorkerConfigLike, WorkerId, WorkerWithDpRank, compute_seq_hash_for_block,
+    ExternalSequenceBlockHash, KvCacheEvent, KvCacheEventData, KvCacheRemoveData, KvCacheStoreData,
+    KvCacheStoredBlockData, LocalBlockHash, OverlapScores, RouterEvent, WorkerConfigLike, WorkerId,
+    WorkerWithDpRank, compute_seq_hash_for_block,
 };
-use crate::sequences::SequencePublisher;
+pub use crate::sequences::NoopSequencePublisher;
 
 pub fn router_event(
     worker_id: WorkerId,
@@ -364,22 +363,6 @@ pub fn assert_overlap_scores_eq(left: &OverlapScores, right: &OverlapScores, ctx
         "frequencies mismatch ({ctx}): left={:?} right={:?}",
         left.frequencies, right.frequencies
     );
-}
-
-/// No-op [`SequencePublisher`] for tests and benchmarks that don't need event transport.
-pub struct NoopSequencePublisher;
-
-impl SequencePublisher for NoopSequencePublisher {
-    fn publish_event(
-        &self,
-        _event: &ActiveSequenceEvent,
-    ) -> impl future::Future<Output = anyhow::Result<()>> + Send {
-        future::ready(Ok(()))
-    }
-
-    fn publish_load(&self, _load: ActiveLoad) {}
-
-    fn observe_load(&self, _: &WorkerWithDpRank, _: &str, _: usize, _: usize) {}
 }
 
 /// Minimal [`WorkerConfigLike`] for scheduler/queue tests and benchmarks.
